Update draft for ruby version of new graph project
This commit:

- Instructs learners to use an adjacency list instead of adjacency
matrix, as adjacency lists offer several benefits over matrices
regarding ease of comprehension, implementation complexity, state
management, and typical-use cases.

- Adds a corresponding explanation on the usage of adjacency lists and
links an additional G4G article that discusses the time and space
complexity trade-offs in the graph operations when using a list vs. a
matrix.

- Changes the `print graph` operation **Tip** to a required function
that learners must implement on their Graph. The G4G article describes
the adjacency list representation as using an array to store the
vertices, and linked lists to store the edges for each element in the
vertices array. Learners may elect to re-use the LinkedList they made in
the previous LinkedList project, or they may elect to use other
(builtin) structures offered by their language. Hard-coding a particular
`print` function forces them into a particular implementation, which
would require us to explicitly instruct the learners to use this
implementation, but this is a choice they should be making so that they
can practice making such design decisions themselves.